+++ .gitignore
@@ -19,7 +19,4 @@ redis-data/
 .vscode/
 
 #temporaire
-scripts/mailjet.py
-documents/HLM-template-v*
 static/icons/ikonate/
-static/highcharts/Highcharts-*
