+++ conventions/services/attribution.py
@@ -1,6 +1,6 @@
 from conventions.forms import ConventionFoyerAttributionForm
 from conventions.services import utils
-from conventions.services.services_conventions import ConventionService
+from conventions.services.conventions import ConventionService
 
 
 class ConventionFoyerAttributionService(ConventionService):
