+++ conventions/services/collectif.py
@@ -1,5 +1,5 @@
 from conventions.forms import UploadForm
-from conventions.services.services_conventions import ConventionService
+from conventions.services.conventions import ConventionService
 from programmes.models import LocauxCollectifs
 from programmes.subforms.lot_collectif import LocauxCollectifsFormSet, LotCollectifForm
