+++ conventions/services/variantes.py
@@ -1,6 +1,6 @@
 from conventions.forms import ConventionFoyerVariantesForm
 from conventions.services import utils
-from conventions.services.services_conventions import ConventionService
+from conventions.services.conventions import ConventionService
 
 
 class ConventionFoyerVariantesService(ConventionService):
