+++ conventions/tests/services/test_annexes_service.py
@@ -7,7 +7,7 @@ from conventions.forms import (
     UploadForm,
 )
 from conventions.models import Convention
-from conventions.services.services_logements import ConventionAnnexesService
+from conventions.services.annexes import ConventionAnnexesService
 from conventions.services import (
     utils,
 )
